+++ src/Main_App/AGENTS.md
@@ -6,5 +6,7 @@ so that it does not become unresponsive.
 
 The eeg_preprocessing.py file should not be edited for any reason. 
 
+post_process.py should not be edited for any reason. 
+
 The goal of this directory is to decrease the size of the core fpvs_app.py file
 while ensuring that future updates are easier by modularizing the code.
